split up retry code
